+++ lib/feedzirra/parser/rss_feed_burner_entry.rb
@@ -16,6 +16,9 @@ module Feedzirra
         element :"content:encoded", :as => :content
         element :description, :as => :summary
 
+        element :"media:content", :as => :image, :value => :url
+        element :enclosure, :as => :image, :value => :url
+
         element :pubDate, :as => :published
         element :pubdate, :as => :published
         element :"dc:date", :as => :published