+++ HISTORY.md
@@ -1,4 +1,7 @@
 # HEAD
+* Use `Time.parse_safely` in `Feed.last_modified_from_header` [[#129](https://github.com/pauldix/feedzirra/pull/129)].
+* Added image to the RSS Entry Parser [[#103](https://github.com/pauldix/feedzirra/pull/103)].
+* Compatibility fixes for Ruby 2.0 [[#136](https://github.com/pauldix/feedzirra/pull/136)].
 * Remove gorillib dependency [[#113](https://github.com/pauldix/feedzirra/pull/113)].
 
 # 0.2.0.rc2